moving dims order to host

+++ cuda/pms/kernels.cu
@@ -14,7 +14,7 @@ __global__ void zero_vector_float(float *vec, const int n){
   if (xIndex < n) vec[xIndex]=0.0f;
 }
 
-__global__ void fill(float *vec, int value, int n){
+__global__ void fill(int *vec, int value, int n){
   unsigned int xIndex = blockDim.x * blockIdx.x + threadIdx.x;
   if (xIndex < n) vec[xIndex]=value;
 }
